settings inputs refactor
