+++ components/SettingsEmailForm/index.js
@@ -1,7 +1,7 @@
 import { useState } from "react";
 
-import { useAuth } from "../../context/AuthContext";
 import UserModel from "../../model/user";
+import { useAuth } from "../../context/AuthContext";
 import {
   validateEmailRegex,
   validateEmailMsg,
@@ -9,25 +9,14 @@ import {
 
 import styles from "./styles.module.css";
 
-const SettingsInputs = ({ user }) => {
-  const { firstName, email, uid } = user;
-  const [newName, setNewName] = useState(firstName);
+const SettingsEmailForm = ({ user }) => {
+  const { email, uid } = user;
   const [newEmail, setNewEmail] = useState(email);
   const [password, setPassword] = useState("");
-  const [isNameSaved, setIsNameSaved] = useState(false);
   const [isEmailSaved, setIsEmailSaved] = useState(false);
   const [showPassword, setShowPassword] = useState(false);
   const { reauthenticate, changeEmail } = useAuth();
 
-  const handleNewName = async (e) => {
-    try {
-      e.preventDefault();
-      await UserModel.updateName(uid, newName);
-      setIsNameSaved(true);
-    } catch (err) {
-      console.log(err);
-    }
-  };
   const handleNewEmail = async (e) => {
     try {
       e.preventDefault();
@@ -43,25 +32,6 @@ const SettingsInputs = ({ user }) => {
 
   return (
     <form className={styles.formContainer} onSubmit={handleNewEmail}>
-      <label htmlFor="text">Name</label>
-      <div className={styles.inputButtonPair}>
-        <input
-          className={styles.input}
-          id="text"
-          type="text"
-          value={newName}
-          onChange={(e) => setNewName(e.target.value)}
-          onFocus={() => setIsNameSaved(false)}
-        />
-        <button
-          className={
-            !isNameSaved ? styles.inputButton : styles.inputButtonSaved
-          }
-          onClick={handleNewName}
-        >
-          {!isNameSaved ? "Save" : "Saved!"}
-        </button>
-      </div>
       <label
         htmlFor="password"
         style={!showPassword ? { display: "none" } : null}
@@ -72,7 +42,7 @@ const SettingsInputs = ({ user }) => {
         <input
           id="password"
           type="password"
-          placeholder="Password"
+          placeholder="Confirm Password"
           autoComplete="current-password"
           value={password}
           className={styles.password}
@@ -112,4 +82,4 @@ const SettingsInputs = ({ user }) => {
   );
 };
 
-export default SettingsInputs;
+export default SettingsEmailForm;
