+++ components/SettingsEmailForm/styles.module.css
@@ -62,12 +62,12 @@
 
   .input {
     height: 95px;
-    width: 450px;
+    width: 435px;
     font-size: x-large;
   }
   .password {
     height: 95px;
-    width: 450px;
+    width: 435px;
     font-size: x-large;
   }
   .inputButton,
