+++ components/SettingsNameForm/styles.module.css
@@ -1,0 +1,61 @@
+.formContainer {
+  display: flex;
+  flex-direction: column;
+  align-self: center;
+  width: 307px;
+}
+
+.inputButtonPair {
+  display: flex;
+  flex-direction: row;
+  gap: 15px;
+  margin-bottom: 24px;
+}
+
+.input {
+  background: none;
+  height: 50px;
+  width: 222px;
+  border-width: 0 0 1px;
+  border-color: black;
+}
+
+.inputButton,
+.inputButtonSaved {
+  background-color: rgba(235, 221, 245, 0.3);
+  width: 60px;
+  border-radius: 6px;
+  border-width: 1px;
+  margin: 10px 0;
+}
+
+.inputButton {
+  border: 1px solid gray;
+}
+.inputButtonSaved {
+  border: 1.8px solid var(--purple-dark);
+}
+
+.container label ~ label {
+  margin-top: 15px;
+}
+
+@media (min-width: 768px) {
+  .formContainer {
+    width: 596px;
+  }
+
+  .input {
+    height: 95px;
+    width: 435px;
+    font-size: x-large;
+  }
+
+  .inputButton,
+  .inputButtonSaved {
+    width: 150px;
+    border-radius: 8px;
+    font-size: x-large;
+    border: 2.8px solid var(--purple-dark);
+  }
+}
