+++ pages/settings/index.js
@@ -1,13 +1,13 @@
 import { useEffect, useState } from "react";
 
 import SettingsTopInfo from "../../components/SettingsTopInfo";
-import SettingsInputs from "../../components/SettingsInputs";
+import SettingsNameForm from "../../components/SettingsNameForm";
+import SettingsEmailForm from "../../components/SettingsEmailForm";
 import SettingsButtons from "../../components/SettingsButtons";
 import Heading1 from "../../components/Heading1";
 import BottomTabs from "../../components/BottomTabs";
 import Loading from "../../components/Loading";
 import UserModel from "../../model/user";
-import { useAuth } from "../../context/AuthContext";
 import withProtected from "../../routers/withProtected";
 
 import styles from "./styles.module.css";
@@ -34,7 +34,8 @@ const SettingsPage = ({ authedUser }) => {
     <div className={styles.container}>
       <Heading1>Settings</Heading1>
       <SettingsTopInfo user={user} />
-      <SettingsInputs user={user} />
+      <SettingsNameForm user={user} />
+      <SettingsEmailForm user={user} />
       <SettingsButtons />
       <BottomTabs />
     </div>
